Converted podcast page to TS
