+++ src/app/podcast/page.tsx
@@ -1,0 +1,247 @@
+'use client';
+import Navbar from '@/components/Navbar';
+import Footer from '@/components/Footer';
+import Hero from '@/components/Hero';
+import Image from 'next/image';
+import Link from 'next/link';
+import { useTheme } from 'next-themes';
+import { useEffect, useState } from 'react';
+import { AnimatePresence, motion } from 'framer-motion';
+import './podcast.css';
+import ScrollToTopButton from '@/components/ScrollToTop';
+import { ListenOn, PodcastQuote } from '@/types/interfaces';
+import PodcastEpisode from '@/components/PodcastEpisode';
+
+export default function PodcastPage() {
+  const { resolvedTheme } = useTheme();
+  const [mounted, setMounted] = useState(false);
+  const [quoteIndex, setQuoteIndex] = useState(0);
+
+  const quotes: PodcastQuote[] = [
+    {
+      text: 'I’m a podcast fanatic – and this is my new obsession! So many life skills tucked in the fun conversations...',
+      author: '— iTunes Review',
+    },
+    {
+      text: "Mark's coaching helped me finally get clarity in my business.",
+      author: '— Entrepreneur Weekly',
+    },
+    {
+      text: 'Every episode delivers pure value. Highly recommend!',
+      author: '— FounderCast',
+    },
+  ];
+
+  const listenOn: ListenOn[] = [
+    {
+      href: new URL('https://www.tiktok.com/@the.phelan.focus'),
+      src: '/podcast/tiktok.png',
+      alt: 'TikTok logo',
+    },
+    {
+      href: new URL('https://www.youtube.com/@ThePhelanFocus'),
+      src: '/podcast/youtube.png',
+      alt: 'YouTube logo',
+    },
+    {
+      href: new URL('https://open.spotify.com/show/1hsQ9TSkvmuzkbnrE27ho7?si=95e611acecfa4385'),
+      src: '/podcast/spotify.png',
+      alt: 'Spotify logo',
+    },
+    {
+      href: new URL('https://www.instagram.com/The.Phelan.Focus/'),
+      src: '/podcast/Instagram.png',
+      alt: 'Instagram logo',
+    },
+  ];
+
+  useEffect(() => {
+    setMounted(true);
+    const interval = setInterval(() => {
+      setQuoteIndex((prev) => (prev + 1) % quotes.length);
+    }, 3000);
+    return () => clearInterval(interval);
+  }, [quotes.length]);
+
+  return (
+    <div className="bg-white dark:bg-[#0b1727] text-gray-800 dark:text-white">
+      <Navbar />
+
+      {/* HERO */}
+      {/* Remove fixed positioning and z-index from Hero */}
+      <div style={{ position: 'relative', width: '100%', height: '50vh', zIndex: 0 }}>
+        <Hero bgImage="/podcast2.jpeg" header="Podcast" />
+      </div>
+
+      {/* MARK PHELAN STORY */}
+      <section
+        className="bg-[#0b1727] py-20 px-6 w-full grid grid-cols-1 lg:grid-cols-2 gap-12 items-center"
+        style={{ position: 'relative', zIndex: 1 }}
+      >
+        <div className="flex justify-center">
+          <Image
+            src="/Phelancoaching.jpg"
+            alt="Podcast Hosts"
+            width={400}
+            height={400}
+            className="rounded-xl shadow-2xl border-4 border-white dark:border-gray-700"
+          />
+        </div>
+        <div className="text-white">
+          <h3 className="text-2xl font-bold mb-4">Meet Mark Phelan</h3>
+          <p className="text-lg leading-relaxed opacity-80">
+            Mark Phelan is a dynamic public speaker and strategic consultant known for transforming
+            complexity into clarity. With a foundation in military logistics and a career spanning
+            project management, e-commerce, IT, and engineering, Mark brings a unique blend of
+            precision, discipline, and forward-thinking innovation.
+          </p>
+          <p className="text-lg leading-relaxed opacity-80 mt-4">
+            He has led high-impact initiatives across multimillion-dollar organizations, building
+            scalable systems and empowering teams to perform at their best. Whether he's on stage or
+            in the boardroom, Mark delivers insights that inspire action and drive measurable
+            results — all while staying grounded in authenticity and service.
+          </p>
+        </div>
+      </section>
+
+      {/* QUOTE CAROUSEL */}
+      <section
+        className="bg-slate-800 text-white py-16 px-6 text-center"
+        style={{ position: 'relative', zIndex: 1 }}
+      >
+        <h3 className="text-3xl font-bold mb-12">What Listeners Say</h3>
+        <div className="relative max-w-3xl mx-auto h-48">
+          <AnimatePresence mode="wait">
+            <motion.div
+              key={quoteIndex}
+              initial={{ opacity: 0, y: 20 }}
+              animate={{ opacity: 1, y: 0 }}
+              exit={{ opacity: 0, y: -20 }}
+              transition={{ duration: 0.6 }}
+              className="absolute w-full"
+            >
+              <div className="bg-slate-700 p-6 rounded-xl shadow-md">
+                <p className="italic text-lg mb-4">“{quotes[quoteIndex].text}”</p>
+                <p className="text-yellow-400 text-sm uppercase font-semibold tracking-widest">
+                  {quotes[quoteIndex].author}
+                </p>
+              </div>
+            </motion.div>
+          </AnimatePresence>
+        </div>
+      </section>
+
+      {/* LISTEN ON */}
+      <section
+        className="bg-[#0b1727] py-20 px-6 text-center"
+        style={{ position: 'relative', zIndex: 1 }}
+      >
+        <h3 className="text-2xl font-semibold mb-10">
+          Listen wherever you find your favorite podcasts:
+        </h3>
+        <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-4 gap-8 items-center justify-items-center">
+          {listenOn.map((place, i) => (
+            <a
+              key={i}
+              href={place.href.toString()}
+              target="_blank"
+              rel="noopener noreferrer"
+              className="transition-transform duration-300 hover:scale-105 hover:shadow-xl rounded-lg"
+            >
+              <Image src={place.src} alt={place.alt} width={200} height={50} />
+            </a>
+          ))}
+        </div>
+      </section>
+      <section
+        className="bg-slate-800 text-white py-20 px-6 text-center"
+        style={{ position: 'relative', zIndex: 1 }}
+      >
+        <h3 className="text-3xl font-bold mb-4">Podcast With Our Featured Guest</h3>
+        <p className="text-lg mb-8 max-w-xl mx-auto">
+          Here is a podcast episode with our featured guest
+        </p>
+        <PodcastEpisode
+          embedSrc={
+            new URL(
+              'https://www.podbean.com/player-v2/?from=embed&i=v3h3u-18f2618-pb&square=1&share=1&download=1&fonts=Verdana&skin=1b1b1b&font-color=auto&rtl=0&logo_link=podcast_page&btn-skin=ff6d00&size=300'
+            )
+          }
+          width="100%"
+          height={300}
+        />
+        <button
+          className={`w-full sm:w-auto px-8 py-3 rounded-lg font-semibold transition duration-300 border-2 ${
+            mounted && resolvedTheme === 'light'
+              ? 'border-yellow-500 text-yellow-600 hover:bg-yellow-500 hover:text-white hover:shadow-md'
+              : 'border-yellow-400 text-yellow-400 hover:bg-yellow-400 hover:text-black hover:shadow-md'
+          }`}
+          onClick={() =>
+            window.open(
+              'https://thejobseekersjourney.podbean.com/e/entrepreneurshipandthemomentuminternshipprogram/',
+              '_blank'
+            )
+          }
+        >
+          Listen to Podcast
+        </button>
+      </section>
+      <section
+        className="bg-[#0b1727] py-20 px-6 text-center"
+        style={{ position: 'relative', zIndex: 1 }}
+      >
+        <h3 className="text-3xl font-semibold mb-10">First TPF Episode</h3>
+        <PodcastEpisode
+          embedSrc={
+            new URL(
+              'https://open.spotify.com/embed/episode/0RFqaTngo8JSXrVBlhIGaV/video?utm_source=generator&t=0'
+            )
+          }
+          allow="autoplay; clipboard-write; encrypted-media; fullscreen; picture-in-picture"
+          width={624}
+          height={351}
+          dataTestid="embed-iframe"
+        />
+        <button
+          className={`w-full sm:w-auto px-8 py-3 rounded-lg font-semibold transition duration-300 border-2 ${
+            mounted && resolvedTheme === 'light'
+              ? 'border-yellow-500 text-yellow-600 hover:bg-yellow-500 hover:text-white hover:shadow-md'
+              : 'border-yellow-400 text-yellow-400 hover:bg-yellow-400 hover:text-black hover:shadow-md'
+          }`}
+          onClick={() =>
+            window.open('https://open.spotify.com/episode/0RFqaTngo8JSXrVBlhIGaV', '_blank')
+          }
+        >
+          All Episodes
+        </button>
+      </section>
+      {/* CTA SECTION */}
+      <section
+        className="bg-slate-800 text-white py-20 px-6 text-center"
+        style={{ position: 'relative', zIndex: 1 }}
+      >
+        <h3 className="text-3xl font-bold mb-4">Are you ready to take action?</h3>
+        <p className="text-lg mb-8 max-w-xl mx-auto">
+          Schedule a no-cost discovery call to learn if I'm the right person to help you move
+          forward.
+        </p>
+        <div className="max-w-xs mx-auto">
+          <Link href="/contact">
+            <button
+              className={`w-full sm:w-auto px-8 py-3 rounded-lg font-semibold transition duration-300 border-2 ${
+                mounted && resolvedTheme === 'light'
+                  ? 'border-yellow-500 text-yellow-600 hover:bg-yellow-500 hover:text-white hover:shadow-md'
+                  : 'border-yellow-400 text-yellow-400 hover:bg-yellow-400 hover:text-black hover:shadow-md'
+              }`}
+            >
+              Book a Call
+            </button>
+          </Link>
+        </div>
+      </section>
+
+      <Footer />
+      <ScrollToTopButton />
+    </div>
+  );
+}
